大量UI优化: 主题即时生效、登录页美化、预览滑动动画、菜单改进、图片缓存修复等

+++ Android/app/src/main/java/com/mrcwoods/memories/UploadTask.java
@@ -10,7 +10,7 @@ public class UploadTask {
 
     public final Uri uri;
     public int status = WAITING;
-    public int progress;
+    public float progress;
     public String message = "等待上传";
     public String url = "";
